app/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
beautifulsoup4>=4.12
pandas
requests
SQLAlchemy>=2.0
aiosqlite>=0.19
pydantic>=2.7